requirements.txt
google-generativeai>=0.8.0
google-genai>=1.0.0
openai>=1.0.0
streamlit>=1.30.0
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.0
Pillow>=10.0.0
reportlab>=4.0.0
plotly>=5.18.0
seaborn>=0.13.0
python-dotenv>=1.0.0
nbformat>=5.9.0
